chore(sdk): Extract Sliding Sync request generators into their own module.
